compiles, currentEvaluator not wired up
