+++ main/resolve/src/mill/resolve/ParseArgs.scala
@@ -1,4 +1,4 @@
-package mill.main
+package mill.resolve
 
 import fastparse.NoWhitespace.noWhitespaceImplicit
 import fastparse._
