+++ main/resolve/test/resources/examples/javac/resources/META-INF/License
@@ -1,0 +1,1 @@
+Dummy Licence
