+++ main/resolve/test/resources/examples/javac/resources/hello.txt
@@ -1,0 +1,1 @@
+Hello World
