+++ main/resolve/test/resources/examples/javac/src/Bar.java
@@ -1,0 +1,4 @@
+package test;
+public class Bar{
+    static int value = 271828;
+}
