+++ main/resolve/test/resources/examples/javac/src/Foo.java
@@ -1,0 +1,7 @@
+package test;
+public class Foo{
+    static int value = 31337;
+    public static void main(String[] args){
+        System.out.println(value + Bar.value);
+    }
+}
