+++ main/testkit/src/mill/testkit/MillTestkit.scala
@@ -8,7 +8,7 @@ import mill.api.Strict.Agg
 
 import java.io.{InputStream, PrintStream}
 import mill.eval.Evaluator
-import mill.main.{ResolveTasks, SelectMode}
+import mill.resolve.{ResolveTasks, SelectMode}
 import mill.util.PrintLogger
 
 import language.experimental.macros
@@ -108,7 +108,7 @@ trait MillTestKit {
     )
 
     def evalTokens(args: String*): Either[Result.Failing[_], (Seq[_], Int)] = {
-      ResolveTasks.resolve(evaluator, args, SelectMode.Separated) match {
+      ResolveTasks.resolve(evaluator.rootModule, args, SelectMode.Separated) match {
         case Left(err) => Left(Result.Failure(err))
         case Right(resolved) => apply(resolved)
       }
