+++ runner/src/mill/runner/MillBuildBootstrap.scala
@@ -3,7 +3,8 @@ import mill.util.{ColorLogger, PrefixLogger, Util, Watchable}
 import mill.{BuildInfo, T}
 import mill.api.{PathRef, Val, internal}
 import mill.eval.Evaluator
-import mill.main.{RootModule, RunScript, SelectMode}
+import mill.main.{RootModule, RunScript}
+import mill.resolve.SelectMode
 import mill.main.TokenReaders._
 import mill.define.{Discover, Segments}
